A principio arrumei a cagada

+++ SiteSample2020/discListar.php
@@ -8,20 +8,6 @@ Abril/2020
 <html>
 <head>
 
-<title>AP - AKIM PNEUS</title>
-<link rel="icon" type="image/png" href="imagens/IE_favicon.png"/>
-<meta name="viewport" content="width=device-width, initial-scale=1">
-<link rel="stylesheet" href="https://www.w3schools.com/w3css/4/w3.css">
-<style>
-    .w3-theme {
-        color: #ffff !important;
-        background-color: #ADD8E6 !important
-    }
-
-    .w3-code {
-        border-left: 4px solid #ADD8E6
-    }
-=======
     <title>AP - AKIM PNEUS</title>
     <link rel="icon" type="image/png" href="imagens/IE_favicon.png"/>
     <meta name="viewport" content="width=device-width, initial-scale=1">
@@ -29,11 +15,11 @@ Abril/2020
     <style>
         .w3-theme {
             color: #ffff !important;
-            background-color: #380077 !important
+            background-color: #ADD8E6 !important
         }
 
         .w3-code {
-            border-left: 4px solid #380077
+            border-left: 4px solid #ADD8E6
         }
 
         .myMenu {
@@ -91,20 +77,6 @@ Abril/2020
 
             // Faz Select na Base de Dados
 
-            $sql = "SELECT CodDisciplina, NomeDisc, Ementa FROM Disciplina";
-            echo "<div class='w3-responsive w3-card-4'>";
-            if ($result = mysqli_query($conn, $sql)) {
-                echo "<table class='w3-table-all'>";
-                echo "	<tr>";
-                echo "	  <th>Código</th>";
-                echo "	  <th width='10%'>Nome</th>";
-				echo "	  <th>Ementa</th>";
-				echo "	  <th> </th>";
-				echo "	  <th> </th>";
-                echo "	</tr>";
-                if (mysqli_num_rows($result) > 0) {
-                	
-
                         $sql = "SELECT CodDisciplina, NomeDisc, Ementa FROM Disciplina";
                         echo "<div class='w3-responsive w3-card-4'>";
                         if ($result = mysqli_query($conn, $sql)) {
@@ -117,34 +89,48 @@ Abril/2020
                             echo "	  <th> </th>";
                             echo "	</tr>";
                             if (mysqli_num_rows($result) > 0) {
+                               
+
+                                $sql = "SELECT CodDisciplina, NomeDisc, Ementa FROM Disciplina";
+                                echo "<div class='w3-responsive w3-card-4'>";
+                                if ($result = mysqli_query($conn, $sql)) {
+                                    echo "<table class='w3-table-all'>";
+                                    echo "	<tr>";
+                                    echo "	  <th>Código</th>";
+                                    echo "	  <th width='10%'>Nome</th>";
+                                    echo "	  <th>Ementa</th>";
+                                    echo "	  <th> </th>";
+                                    echo "	  <th> </th>";
+                                    echo "	</tr>";
+                                    if (mysqli_num_rows($result) > 0) {
 
                     // Apresenta cada linha da tabela
-                                while ($row = mysqli_fetch_assoc($result)) {
-                                    $cod = $row["CodDisciplina"];
-                                    echo "<tr>";
-                                    echo "<td>";
-                                    echo $cod;
-                                    echo "</td><td>";
-                                    echo $row["NomeDisc"];
-                                    echo "</td><td>";
-                                    echo $row["Ementa"];
-                                    echo "</td><td>";
+                                        while ($row = mysqli_fetch_assoc($result)) {
+                                            $cod = $row["CodDisciplina"];
+                                            echo "<tr>";
+                                            echo "<td>";
+                                            echo $cod;
+                                            echo "</td><td>";
+                                            echo $row["NomeDisc"];
+                                            echo "</td><td>";
+                                            echo $row["Ementa"];
+                                            echo "</td><td>";
 
 						//Atualizar e Excluir registro de prof
-                                    ?>
-                                    <a href='discAtualizar.php?id=<?php echo $cod; ?>'><img src='imagens/Edit.png' title='Editar Disciplina' width='32'></a>
-                                </td><td>
-                                    <a href='discExcluir.php?id=<?php echo $cod; ?>'><img src='imagens/Delete.png' title='Excluir Disciplina' width='32'></a>
-                                </td>
-                            </tr>
-                            <?php
+                                            ?>
+                                            <a href='discAtualizar.php?id=<?php echo $cod; ?>'><img src='imagens/Edit.png' title='Editar Disciplina' width='32'></a>
+                                        </td><td>
+                                            <a href='discExcluir.php?id=<?php echo $cod; ?>'><img src='imagens/Delete.png' title='Excluir Disciplina' width='32'></a>
+                                        </td>
+                                    </tr>
+                                    <?php
+                                }
+                            }
+                            echo "</table>";
+                            echo "</div>";
+                        } else {
+                            echo "Erro executando SELECT: " . mysqli_error($conn);
                         }
-                    }
-                    echo "</table>";
-                    echo "</div>";
-                } else {
-                    echo "Erro executando SELECT: " . mysqli_error($conn);
-                }
 
             mysqli_close($conn);  //Encerra conexao com o BD
 
@@ -167,4 +153,4 @@ Abril/2020
 <!-- Inclui RODAPE.PHP  -->
 <?php require 'rodape.php';?>
 </body>
-</html>
+</html
